+++ src/pages/BackOffice/Test.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Test = () => {
+  return <div>Test</div>;
+};
+
+export default Test;
